+++ lib/src/televerse/sessions/sessions.dart
@@ -53,6 +53,56 @@ part of televerse;
 /// }
 /// ```
 abstract class Session {
+  /// Saves the session to a file.
+  ///
+  /// By default, the sessions are saved in the `.televerse/sessions` directory. If you want to change the directory, you can use the `path` parameter.
+  /// The path should be a directory path. The file name will be the session id.
+  /// ```dart
+  /// bot.sessions.saveAllSessions(path: '/path/to/sessions');
+  /// ```
+  void saveToFile({String? path}) {
+    io.Directory directory = io.Directory.current;
+    io.Directory sessionDir =
+        io.Directory(path ?? "${directory.path}/.televerse/sessions");
+
+    if (!sessionDir.existsSync()) {
+      sessionDir.createSync(recursive: true);
+    }
+
+    if (path == null && id == null) {
+      throw TeleverseException.sessionIdNotSet;
+    }
+
+    final filePath = "${sessionDir.path}/$id.json";
+
+    final json = JsonEncoder.withIndent('  ').convert(toJson());
+    io.File(filePath).writeAsStringSync(json);
+  }
+
+  /// Loads the session from a file.
+  ///
+  /// Returns null if the file does not exist.
+  static T? loadFromFile<T extends Session>(
+    T Function(Map<String, dynamic> json) fromJson, {
+    String? path,
+    int? id,
+  }) {
+    if (path == null && id == null) {
+      throw TeleverseException.sessionIdNotSet;
+    }
+
+    path ??= "${io.Directory.current.path}/.televerse/sessions/$id.json";
+
+    io.File file = io.File(path);
+
+    if (!file.existsSync()) {
+      return null;
+    }
+
+    final json = file.readAsStringSync();
+    return fromJson(jsonDecode(json));
+  }
+
   /// The session id - this is the chat id.
   int? id;
 
@@ -67,7 +117,37 @@ abstract class Session {
 }
 
 /// Manages the sessions for Televerse.
+/// Session manger instance will be available as `bot.sessions`.
+///
+/// Session manager can help you to save all the sessions to their respective files.
+///
+/// ```dart
+/// bot.sessions.saveAllSessions();
+/// ```
+///
+/// By default, the sessions are saved in the `.televerse/sessions` directory. If you want to change the directory, you can use the `path` parameter.
+/// The path should be a directory path. The file name will be the session id.
+/// ```dart
+/// bot.sessions.saveAllSessions(path: '/path/to/sessions');
+/// ```
 class SessionsManager<T extends Session> {
+  /// Saves all the sessions to their respective files.
+  ///
+  /// ```dart
+  /// bot.sessions.saveAllSessions();
+  /// ```
+  ///
+  /// By default, the sessions are saved in the `.televerse/sessions` directory. If you want to change the directory, you can use the `path` parameter.
+  /// The path should be a directory path. The file name will be the session id.
+  /// ```dart
+  /// bot.sessions.saveAllSessions(path: '/path/to/sessions');
+  /// ```
+  void saveAllSessions({String? path}) {
+    for (final session in _sessions.values) {
+      session?.saveToFile(path: path);
+    }
+  }
+
   /// Constant enabled
   bool enabled = true;